Purpose of this project.
This is a practice exercise. It's purpose is to practice putting HTML, CSS and Flexbox together in a project. It is a deviation from the Odin Project's normal curriculum, but I feel as though the excercise is necessary to further cement the use of all three programs. 

I will be taking a look at the layout of example websites found on articles about just this topic; this will serve as the base inspiration, but the individual details will be left to me